keep crypto/rand as the default
